clean up bazaar/order and bazaar/price packages
