+++ src/main/java/com/github/mkram17/bazaarutils/utils/bazaar/market/price/PriceInfo.java
@@ -1,28 +1,25 @@
 package com.github.mkram17.bazaarutils.utils.bazaar.market.price;
 
-import com.github.mkram17.bazaarutils.utils.bazaar.data.BazaarDataManager;
-import com.github.mkram17.bazaarutils.utils.Util;
 import com.github.mkram17.bazaarutils.utils.bazaar.market.order.Order;
 import com.github.mkram17.bazaarutils.utils.bazaar.market.order.OrderInfo;
 import com.github.mkram17.bazaarutils.utils.bazaar.market.order.OrderType;
-import com.teamresourceful.resourcefulconfig.api.annotations.ConfigEntry;
-import com.teamresourceful.resourcefulconfig.api.annotations.ConfigObject;
 import lombok.Getter;
 import lombok.Setter;
+import lombok.ToString;
 
 /**
  * Container for market price metadata of a single product. For actual user orders, prefer
  * {@link OrderInfo} or {@link Order}.
  */
-@ConfigObject
+@ToString
 public class PriceInfo {
-    @Setter @Getter @ConfigEntry(id = "orderType")
+    @Setter @Getter
     protected OrderType orderType;
 
-    @Setter @Getter @ConfigEntry(id = "pricingPosition")
+    @Setter @Getter
     protected PricingPosition pricingPosition;
 
-    @Setter @Getter @ConfigEntry(id = "pricePerItem")
+    @Setter @Getter
     protected Double pricePerItem;
 
 
@@ -38,10 +35,5 @@ public class PriceInfo {
             //TODO revisit whether this still needs to have default value
             this.orderType = OrderType.SELL;
         }
-    }
-
-    public void flipPrices(double newPrice) {
-        this.orderType = this.orderType.getOpposite();
-        this.pricePerItem = newPrice;
     }
 }
